num is digit (list)
Create a window that prompts the user to enter a number in a text field. When the button is clicked, the varialble.isdigit() construct is used to check if the number is an integer. If the check is positive, it is added to the list; otherwise, the contents of the text field are erased.