tkinter (csv) 2
Creates a program that allows the user to create a new .csv file. The program prompts you to enter a name and age, and then adds the entered data to the end of the newly created file + created a button to display the contents of a .csv file by importing its data into a list.