tkinter sum
A program that generates two random integers between 10 and 50. It prompts the user to add the numbers and enter an answer. If the user gives the correct answer, the appropriate image will be displayed; if the answer is wrong, display another suitable image. To get another question, the user must click the Next button